requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5

# Environment and Configuration
python-decouple==3.8

# Database
PyMySQL==1.1.0
sshtunnel==0.4.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Caching (optional but recommended)
django-redis==5.4.0
redis==5.0.1

# Static Files
whitenoise==6.6.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Async Support (for scraper optimization)
celery==5.3.4
django-celery-beat==2.5.0

# Monitoring and Logging
sentry-sdk==1.38.0

# API Testing
httpie==3.2.2